requirements.txt
aiohttp==1.0.5
async-timeout==3.0.1
certifi==2019.3.9
chardet==3.0.4
discord==0.16.12
discord.py==0.16.12
idna==2.8
multidict==4.5.2
Pillow==5.4.1
requests==2.21.0
urllib3==1.24.1
websockets==3.4
